requirements.txt
pyrogram>=2.0.106
TgCrypto>=1.2.5
httpx>=0.27.0
structlog>=24.1.0
python-dotenv>=1.0.0
psutil>=6.0.0
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.5
ruff>=0.3.0
edge-tts>=6.1.9
pydub>=0.25.1
playwright>=1.41.0
mcp>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

Pillow>=10.0.0
pillow-heif>=0.15.0
google-genai>=1.0.0
mlx-whisper>=0.0.1
pyyaml>=6.0
structlog>=24.1.0
mss
PyPDF2>=3.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
aiofiles>=23.2.1
APScheduler>=3.10.4
dateparser>=1.2.0
sounddevice
